+++ shared/login/signup/request-invite-success/container.js
@@ -1,7 +1,7 @@
 // @flow
-import * as Creators from '../../actions/signup'
-import Render from './request-invite-success.render'
-import {connect} from '../../util/container'
+import * as Creators from '../../../actions/signup'
+import Render from '.'
+import {connect} from '../../../util/container'
 
 const mapStateToProps = state => ({})
 const mapDispatchToProps = (dispatch: Dispatch) => ({
