+++ shared/login/signup/request-invite/index.desktop.js
@@ -1,9 +1,9 @@
 // @flow
-import Container from '../forms/container'
+import Container from '../../forms/container'
 import React, {Component} from 'react'
-import type {Props} from './request-invite.render'
-import {Box, Text, Icon, Input, Button} from '../../common-adapters'
-import {globalStyles, globalMargins} from '../../styles'
+import {Box, Text, Icon, Input, Button} from '../../../common-adapters'
+import {globalStyles, globalMargins} from '../../../styles'
+import type {Props} from '.'
 
 class RequestInviteRender extends Component<Props> {
   render() {
