+++ shared/login/signup/username-email-form/index.desktop.js
@@ -1,37 +1,32 @@
 // @flow
-import Container from '../forms/container'
+import Container from '../../forms/container'
 import React, {Component} from 'react'
-import type {Props} from './username-email-form.render'
-import {UserCard, Input, Button} from '../../common-adapters'
-import {globalColors, globalMargins} from '../../styles'
+import {UserCard, Input, Button} from '../../../common-adapters'
+import {globalStyles, globalColors} from '../../../styles'
+import type {Props} from '.'
 
 class UsernameEmailFormRender extends Component<Props> {
   render() {
     return (
-      <Container onBack={this.props.onBack} outerStyle={stylesOuter}>
-        <UserCard>
+      <Container onBack={this.props.onBack} style={stylesContainer} outerStyle={stylesOuter}>
+        <UserCard style={stylesCard}>
           <Input
             autoFocus={true}
             hintText="Create a username"
-            floatingHintTextOverride="Create a username"
             value={this.props.username}
             errorText={this.props.usernameErrorText}
             onChangeText={username => this.props.usernameChange(username)}
           />
           <Input
-            style={stylesInput2}
             hintText="Email address"
-            floatingHintTextOverride="Email address"
             value={this.props.email}
             errorText={this.props.emailErrorText}
             onEnterKeyDown={this.props.onSubmit}
             onChangeText={email => this.props.emailChange(email)}
-            keyboardType="email-address"
           />
           <Button
             waiting={this.props.waiting}
-            style={stylesButton}
-            fullWidth={true}
+            style={{marginTop: 40, alignSelf: 'center'}}
             type="Primary"
             label="Continue"
             onClick={this.props.onSubmit}
@@ -45,13 +40,14 @@ class UsernameEmailFormRender extends Component<Props> {
 const stylesOuter = {
   backgroundColor: globalColors.white,
 }
-
-const stylesInput2 = {
-  marginTop: globalMargins.tiny,
+const stylesContainer = {
+  ...globalStyles.flexBoxColumn,
+  justifyContent: 'center',
+  alignItems: 'center',
+  marginTop: 15,
 }
-
-const stylesButton = {
-  marginTop: globalMargins.small,
+const stylesCard = {
+  alignItems: 'stretch',
 }
 
 export default UsernameEmailFormRender
